risch: hackish, but more-real fix to lots of recent problems
After Function -> Function2 transition risch coudn't even integrate elementary
functions.

The problem seems to be related to the fact that integration variable sometimes
is the first (maybe just not last) item in terms.

I simply move it to the end of terms and this tweak appers to solve a lot of
problems.

I've also killed explicit integral tables crated by Ondrej -- they are not
needed now.

No need for new tests -- existing tests catch the bug.